# Optimize cursor representations (making history lazy, etc.)