There are only minor changes to the API here (e.g. no more `HCursor.apply`), but `HCursor` and `ACursor` now use slightly different representations, and the history provided by `HCursor` is lazy—the list of `HistoryOp` items won't be created unless you ask for it.

In general you only need the history in the case of a decoding failure, so computing it only on demand seems perfectly reasonable.

This has a pretty huge impact on performance:

Current master:
```
DecodingBenchmark.decodeFoosC      thrpt   80    2175.136 ±   12.538  ops/s
DecodingBenchmark.decodeIntsC      thrpt   80   11732.675 ±  417.534  ops/s
```
And this PR:

```
DecodingBenchmark.decodeFoosC      thrpt   80    2328.876 ±   27.059  ops/s
DecodingBenchmark.decodeIntsC      thrpt   80   13882.966 ±  201.895  ops/s
```

Including allocation rates. For current master:

```
DecodingBenchmark.decodeFoosC:gc.alloc.rate.norm  thrpt   10  1935616.225 ±      0.018    B/op
DecodingBenchmark.decodeIntsC:gc.alloc.rate.norm  thrpt   10   366634.179 ±     10.229    B/op
```

And this PR:

```
DecodingBenchmark.decodeFoosC:gc.alloc.rate.norm  thrpt   10  1712028.188 ±   1306.782    B/op
DecodingBenchmark.decodeIntsC:gc.alloc.rate.norm  thrpt   10   338540.037 ±  19142.773    B/op
```